+++ R/internal-expectedVersion.R
@@ -1,6 +1,6 @@
 #' Expected version
 #'
-#' @note Updated 2021-01-06.
+#' @note Updated 2021-02-11.
 #' @noRd
 #'
 #' @param x `character`.
@@ -24,10 +24,16 @@
         FUN = function(x) {
             x <- syntactic::kebabCase(x)
             keep <- grepl(pattern = paste0("^", x, "="), x = variables)
-            assert(sum(keep, na.rm = TRUE) == 1L)
+            if (!identical(sum(keep, na.rm = TRUE), 1L)) {
+                stop(sprintf("Need to fix koopa variable: '%s'.", x))
+            }
             x <- variables[keep]
             assert(isString(x))
-            x <- sub(pattern = "^(.+)=\"(.+)\"$", replacement = "\\2", x = x)
+            x <- sub(
+                pattern = "^(.+)=\"(.+)\"(.+)?$",
+                replacement = "\\2",
+                x = x
+            )
             x
         },
         FUN.VALUE = character(1L),
@@ -37,18 +43,21 @@
 
 
 
+## Updated 2021-02-11.
 .expectedHomebrewCaskVersion <- function(x) {
     .expectedVersion(paste0("homebrew-cask-", x))
 }
 
 
 
+## Updated 2021-02-11.
 .expectedMacOSAppVersion <- function(x) {
     .expectedVersion(x = paste0("macos-app-", tolower(x)))
 }
 
 
 
+## Updated 2021-02-11.
 .expectedMajorVersion <- function(x) {
     x <- .expectedVersion(x)
     x <- sanitizeVersion(x)
@@ -58,9 +67,31 @@
 
 
 
+## Updated 2021-02-11.
 .expectedMinorVersion <- function(x) {
     x <- .expectedVersion(x)
     x <- sanitizeVersion(x)
     x <- minorVersion(x)
     x
+}
+
+
+
+## Updated 2021-02-11.
+.expectedPythonPackageVersion <- function(x) {
+    .expectedVersion(x = paste("python", x, sep = "-"))
+}
+
+
+
+## Updated 2021-02-11.
+.expectedRubyPackageVersion <- function(x) {
+    .expectedVersion(x = paste("ruby", x, sep = "-"))
+}
+
+
+
+## Updated 2021-02-11.
+.expectedRustPackageVersion <- function(x) {
+    .expectedVersion(x = paste("rust", x, sep = "-"))
 }