+++ R/package.R
@@ -1,5 +1,12 @@
-#' @importFrom AcidBase basenameSansExt fileExt initDir majorVersion
-#'   minorVersion pasteURL realpath requireNamespaces sanitizeVersion shell
+#' koopa
+#'
+#' Shell bootloader for bioinformatics.
+#'
+#' @aliases NULL
+#'
+#' @importFrom AcidBase basenameSansExt fileExt head initDir install.packages
+#'   majorVersion minorVersion pasteURL realpath requireNamespaces
+#'   sanitizeVersion shell tail
 #' @importFrom AcidCLI alert alertInfo alertSuccess alertWarning green h1 h2 h3
 #'   h4 h5 h6 h7 magenta red ul yellow
 #' @importFrom AcidDevTools check install installRecommendedPackages
@@ -10,5 +17,4 @@
 #'   isPackageVersion isString isSubset isSystemCommand
 #' @importFrom methods is
 #' @importFrom syntactic autopadZeros camelCase capitalize kebabCase snakeCase
-#' @importFrom utils install.packages tail
-NULL
+"_PACKAGE"